+++ src/Types/auth.ts
@@ -1,0 +1,5 @@
+export interface User {
+  uid: string;
+  dailyCredits: number;
+  lastCreditRefresh: Date;
+}
