+++ src/Types/tarotCard.ts
